# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexusPay Platform API
  version: 1.0.0
  description: |
    NexusPay is a modern payment platform providing wallet, card, and transfer capabilities.

    All monetary amounts are expressed in integer minor units (kobo for NGN, cents for USD).

    **Authentication**: Most endpoints require a Bearer JWT token obtained via the `/v1/auth/login` or `/v1/auth/register` flows.

    **Idempotency**: Payment creation requests require an `Idempotency-Key` header to prevent duplicate transactions.

    **Error Format**: All errors follow a standard shape:
    ```json
    { "error": "ERROR_CODE", "message": "Human-readable message", "requestId": "optional-trace-id" }
    ```
  contact:
    name: NexusPay Support
    email: support@nexuspay.dev
    url: https://nexuspay.dev
  license:
    name: Proprietary
    url: https://nexuspay.dev/license
servers:
  - url: http://localhost:4000
    description: Local development (API Gateway)
...
paths:
...
  /v1/auth/register:
    post:
      operationId: registerUser
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new user account and returns tokens for immediate login.
        Password must be at least 10 characters and contain uppercase, lowercase, and a digit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: alice@example.com
              password: Secure1234!
              fullName: Alice Johnson
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                user:
                  id: usr_a1b2c3d4
                  email: alice@example.com
                  fullName: Alice Johnson
                  status: ACTIVE
                  isVerified: false
                  mfaEnabled: false
                  createdAt: "2026-08-25T10:00:00Z"
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: rt_1a2b3c4d5e6f7g8h9i0jklmnopqrstuvwx
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/auth/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Authenticate a user
      description: |
        Returns access and refresh tokens on success.
        If MFA is enabled the response contains a `challengeToken` instead of tokens,
        and the caller must POST to `/v1/auth/mfa/verify`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: alice@example.com
              password: Secure1234!
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthResponse"
                  - $ref: "#/components/schemas/MfaChallenge"
              examples:
                success:
                  summary: Direct login (no MFA)
                  value:
                    user:
                      id: usr_a1b2c3d4
                      email: alice@example.com
                      fullName: Alice Johnson
                      status: ACTIVE
                      isVerified: true
                      mfaEnabled: false
                      createdAt: "2026-08-25T10:00:00Z"
                    accessToken: eyJhbGciOiJIUzI1NiJ9...
                    refreshToken: rt_abcdef1234567890abcdef1234567890
                mfaRequired:
                  summary: MFA challenge issued
                  value:
                    mfaRequired: true
                    challengeToken: ct_01H8X9Y7Z6W5V4U3T2S1R0Q9P8O7N6M5L4K3J2H1G0F
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Exchange refresh token for new token pair
      description: Returns a new access and refresh token. The previous refresh token is invalidated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refreshToken: rt_abcdef1234567890abcdef1234567890
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
              example:
                accessToken: eyJhbGciOiJIUzI1NiJ9...
                refreshToken: rt_zyxwvu0987654321zyxwvu0987654321
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/mfa/verify:
    post:
      operationId: verifyMfa
      tags: [Auth]
      summary: Verify MFA code
      description: Completes the MFA login flow by exchanging a challenge token and TOTP code for tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerifyRequest"
            example:
              challengeToken: ct_01H8X9Y7Z6W5V4U3T2S1R0Q9P8O7N6M5L4K3J2H1G0F
              code: "482916"
      responses:
        "200":
          description: MFA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                user:
                  id: usr_a1b2c3d4
                  email: alice@example.com
                  fullName: Alice Johnson
                  status: ACTIVE
                  isVerified: true
                  mfaEnabled: true
                  createdAt: "2026-08-25T10:00:00Z"
                accessToken: eyJhbGciOiJIUzI1NiJ9...
                refreshToken: rt_abcdef1234567890abcdef1234567890
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      operationId: logoutUser
      tags: [Auth]
      summary: Invalidate refresh token
      description: Revokes the given refresh token (or current session if omitted). Returns 204 on success.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              refreshToken: rt_abcdef1234567890abcdef1234567890
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── MFA (authenticated) ─────────────────────────────────────────────
...
  /v1/notifications:
    post:
      operationId: createNotification
      tags: [Notifications]
      summary: Send a notification
      description: |
        Dispatches a notification via the specified channel. The gateway injects the
        `X-Internal-Api-Key` header when forwarding to the notifications service.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
            example:
              userId: usr_a1b2c3d4
              channel: EMAIL
              templateKey: payment.received
              recipient: alice@example.com
              payload:
                amount: "50.00"
                currency: NGN
      responses:
        "202":
          description: Notification accepted for delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationAccepted"
              example:
                id: notif_xyz789
                status: QUEUED
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listNotifications
      tags: [Notifications]
      summary: List notifications
      description: Paginated list of notifications for the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
        - name: cursor
          in: query
          schema:
            type: string
          example: notif_xyz789
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
              example:
                notifications:
                  - id: notif_xyz789
                    channel: EMAIL
                    templateKey: payment.received
                    recipient: alice@example.com
                    status: DELIVERED
                    payload:
                      amount: "50.00"
                      currency: NGN
                    createdAt: "2026-08-25T10:05:00Z"
                    attempts:
                      - status: DELIVERED
                        attemptedAt: "2026-08-25T10:05:01Z"
                nextCursor: null
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/notifications/{id}:
    get:
      operationId: getNotification
      tags: [Notifications]
      summary: Get notification details
      description: Returns a single notification with delivery attempt history.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: notif_xyz789
      responses:
        "200":
          description: Notification details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationDetail"
              example:
                id: notif_xyz789
                channel: EMAIL
                templateKey: payment.received
                recipient: alice@example.com
                status: DELIVERED
                payload:
                  amount: "50.00"
                  currency: NGN
                createdAt: "2026-08-25T10:05:00Z"
                attempts:
                  - status: DELIVERED
                    attemptedAt: "2026-08-25T10:05:01Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Gateway Aggregation ─────────────────────────────────────────────
...
  /v1/mfa/verify:
    post:
      operationId: confirmMfaSetup
      tags: [Auth]
      summary: Confirm MFA setup
      description: Verifies a 6-digit code against the generated secret to complete MFA enrollment.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: MFA enabled
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Email verification ─────────────────────────────────────────────
  /v1/auth/verify-email:
    get:
      operationId: verifyEmail
      tags: [Auth]
      summary: Verify email address
      description: |
        Verifies a user's email address. Currently a stub pending integration
        with notifications-service.
      parameters:
        - name: token
          in: query
          schema:
            type: string
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Templates ──────────────────────────────────────────────────────
  /v1/templates:
    get:
      operationId: listTemplates
      tags: [Notifications]
      summary: List notification templates
      description: Returns all notification templates. Internal API key required.
      security:
        - internalApiKey: []
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: upsertTemplate
      tags: [Notifications]
      summary: Create or update a template
      description: Upserts a notification template by key. Internal API key required.
      security:
        - internalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, channel, subject, body]
              properties:
                key:
                  type: string
                  maxLength: 64
                channel:
                  type: string
                  enum: [EMAIL, SMS]
                subject:
                  type: string
                  maxLength: 200
                body:
                  type: string
      responses:
        "200":
          description: Template saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/templates/{key}:
    get:
      operationId: getTemplate
      tags: [Notifications]
      summary: Get a template by key
      description: Returns a single notification template. Internal API key required.
      security:
        - internalApiKey: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Reconciliation ─────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /v1/auth/login or /v1/auth/register
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
      description: Internal microservice API key for service-to-service calls

  # ── Reusable Parameters ─────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: VALIDATION_ERROR
            message: "Invalid request body: email must be a valid email address"
            requestId: req_abc123
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: "Invalid or expired access token"
            requestId: req_abc123
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: "Payment not found"
            requestId: req_abc123
    Conflict:
      description: Resource state conflict (e.g. already captured, duplicate key)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: CONFLICT
            message: "Payment has already been captured"
            requestId: req_abc123
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 10
          description: Password with min 10 chars, must include uppercase, lowercase, and digit
        fullName:
          type: string
          minLength: 1
          maxLength: 120
          description: Optional display name
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 20
          description: The refresh token to exchange
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Optional refresh token to revoke
    MfaChallenge:
      type: object
      required: [mfaRequired, challengeToken]
      properties:
        mfaRequired:
          type: boolean
          enum: [true]
        challengeToken:
          type: string
          minLength: 20
          description: Ephemeral token for the MFA verification step
    MfaVerifyRequest:
      type: object
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
          minLength: 20
        code:
          type: string
          pattern: "^\\d{6}$"
          description: 6-digit TOTP code
...
    TokenPair:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Opaque refresh token
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          $ref: "#/components/schemas/UserSummary_Small"
        accessToken:
          type: string
        refreshToken:
          type: string
    UserSummary_Small:
      type: object
      required: [id, email, fullName, status, isVerified, mfaEnabled, createdAt]
      properties:
        id:
          type: string
          example: usr_a1b2c3d4
        email:
          type: string
          format: email
        fullName:
          type: string
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DEACTIVATED]
        isVerified:
          type: boolean
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Users ───────────────────────────────────────────────────────
...
    CreateNotificationRequest:
      type: object
      required: [channel, templateKey, recipient, payload]
      properties:
        userId:
          type: string
          description: Optional user ID for association
        channel:
          type: string
          enum: [EMAIL, SMS]
        templateKey:
          type: string
          description: Template identifier (e.g. "payment.received")
        recipient:
          type: string
          description: Email address or phone number
        payload:
          type: object
          additionalProperties: true
          description: Template variables
          example:
            amount: "50.00"
            currency: NGN
    NotificationAccepted:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [QUEUED]
    NotificationSummary:
      type: object
      required: [id, channel, templateKey, recipient, status, payload, createdAt]
      properties:
        id:
          type: string
        channel:
          type: string
          enum: [EMAIL, SMS]
        templateKey:
          type: string
        recipient:
          type: string
        status:
          type: string
          enum: [QUEUED, SENT, DELIVERED, FAILED]
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    NotificationDetail:
      allOf:
        - $ref: "#/components/schemas/NotificationSummary"
        - type: object
          properties:
            attempts:
              type: array
              items:
                $ref: "#/components/schemas/NotificationAttempt"
    NotificationAttempt:
      type: object
      required: [status, attemptedAt]
      properties:
        status:
          type: string
          enum: [SENT, DELIVERED, FAILED]
        attemptedAt:
          type: string
          format: date-time
    NotificationList:
      type: object
      required: [notifications, nextCursor]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationSummary"
        nextCursor:
          type: string
          nullable: true

    # ── Webhooks ────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request body"
        requestId:
          type: string
          nullable: true
          description: Request trace ID for debugging

    # ── Users (admin) ────────────────────────────────────────────────
...
    Template:
      type: object
      required: [key, channel, subject, body]
      properties:
        key:
          type: string
        channel:
          type: string
          enum: [EMAIL, SMS]
        subject:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Reconciliation ───────────────────────────────────────────────
...